Make random port allocation internal
