+++ core/src/main/kotlin/net/corda/core/CordaInternal.kt
@@ -6,6 +6,6 @@ package net.corda.core
  * These fields are only meant to be used by Corda internally, and are not intended to be part of the public API.
  */
 @Retention(AnnotationRetention.BINARY)
-@Target(AnnotationTarget.PROPERTY_GETTER, AnnotationTarget.PROPERTY_SETTER, AnnotationTarget.CLASS)
+@Target(AnnotationTarget.PROPERTY_GETTER, AnnotationTarget.PROPERTY_SETTER)
 @MustBeDocumented
 annotation class CordaInternal
