+++ node/src/test/kotlin/net/corda/node/services/rpc/ArtemisRpcTests.kt
@@ -19,6 +19,7 @@ import net.corda.nodeapi.internal.config.SSLConfiguration
 import net.corda.nodeapi.internal.config.User
 import net.corda.testing.core.SerializationEnvironmentRule
 import net.corda.testing.driver.PortAllocation
+import net.corda.testing.driver.internal.RandomFree
 import org.apache.activemq.artemis.api.core.ActiveMQConnectionTimedOutException
 import org.apache.activemq.artemis.api.core.ActiveMQNotConnectedException
 import org.apache.activemq.artemis.api.core.management.ActiveMQServerControl
@@ -31,7 +32,7 @@ import java.nio.file.Path
 import kotlin.reflect.KClass
 
 class ArtemisRpcTests {
-    private val ports: PortAllocation = PortAllocation.RandomFree
+    private val ports: PortAllocation = RandomFree
 
     private val user = User("mark", "dadada", setOf(all()))
     private val users = listOf(user)
