+++ testing/node-driver/src/integration-test/kotlin/net/corda/testing/driver/DriverTests.kt
@@ -17,6 +17,7 @@ import net.corda.testing.node.internal.internalDriver
 import net.corda.testing.core.DUMMY_BANK_A_NAME
 import net.corda.testing.core.DUMMY_BANK_B_NAME
 import net.corda.testing.core.DUMMY_NOTARY_NAME
+import net.corda.testing.driver.internal.RandomFree
 import net.corda.testing.http.HttpApi
 import net.corda.testing.node.NotarySpec
 import org.assertj.core.api.Assertions.*
@@ -71,7 +72,7 @@ class DriverTests {
 
     @Test
     fun `random free port allocation`() {
-        val nodeHandle = driver(DriverParameters(portAllocation = PortAllocation.RandomFree)) {
+        val nodeHandle = driver(DriverParameters(portAllocation = RandomFree)) {
             val nodeInfo = startNode(providedName = DUMMY_BANK_A_NAME)
             nodeMustBeUp(nodeInfo)
         }
